Second Video Completed
I had trouble fidning the obsolete navigation window, considering the project didnt have the ai navigation package in, so those were a frutrating 15 minutes.
aside from that, it was much quicker than the previous one, I'm content with the result, and I now understand a few things about how you need to connect the changes of the animator to the object, and that they are separate, so thats a caution to have.
well, it was fun